Add GetEmplyoeeInfo api
